refactor: extract changed file helpers
