+++ scripts/utils/index.ts
@@ -2,3 +2,5 @@ export * from './http';
 export * from './platform';
 export * from './frontMatter';
 export * from './postMap';
+export * from './content';
+export * from './git';
